+++ container/src/App.jsx
@@ -1,22 +1,23 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import Container from "./Container";
-import { HashRouter, Route, Switch } from 'react-router-dom';
-import remoteRoutes from 'calender/routes';
+import { HashRouter, Route, Switch } from "react-router-dom";
+import remoteRoutes from "calender/routes";
 import "./index.css";
 import CalenderApp from "./components/CalenderApp";
 import PdpApp from "./components/PdpApp";
-
+import Home from "./components/Home";
 const routes = [...remoteRoutes];
 
 const App = () => (
   <div className="container">
-    <Container/>
+    <Container />
 
     <HashRouter>
       <div>
-      <CalenderApp/>
-      <PdpApp/>
+        <Home />
+        <CalenderApp />
+        <PdpApp />
 
         {/* <React.Suspense fallback={<div>Loading...</div>}>
           <Switch>
@@ -36,6 +37,3 @@ const App = () => (
 );
 
 export default App;
-
-
-
