+++ container/src/components/Home.js
@@ -1,0 +1,11 @@
+import { mount } from "home/home";
+import React, { useRef, useEffect } from "react";
+export default () => {
+  const ref = useRef(null);
+
+  useEffect(() => {
+    mount(ref.current);
+  });
+
+  return <div ref={ref} />;
+};
